The undo button.
Added an undo button that can undo any action the user takes before
submitting an end turn move. The undo button is disabled when there is
no move to undo or during an attack.

The move constructor now defaults to None for move list and build type
to make constructing moves more intuitive.

Pause on start is disable for games with human players because that
feels janky.