Fixed quirk that gave AIs a diff CWD for __init__
From Ryan Regier (Spring 2020):
I was having strange behavior loading our utilities file for HW6.. The file it
would read from in __init__ was different from the file it would write to later,
despite having identical paths.. On init, it would look for the file in the AI/
folder, but on write it would put it in the src/ directory (aka the current
working directory, where the file is required to be read/written from in this
assignment).. It appears the culprit is in the Antics source code itself,
specifically in the loadAIs function of Game.py (starting at line
1144).. Specifically, on line 1154 the current working directory is changed to
the AI/ folder, then all agents are initialized, then on line 1186 the working
directory is changed back.. I'm not sure why this is how the code functions, but
I wanted to point out this inconsistency regardless.. Because the file cannot be
read during __init__, we have to wait to read the file until the first time
getPlacement is called, which feels like a hack to me.

I fixed this by moving the os.chdir('..') call before the agents are loaded.
I hope I didn't break anything